added No profile screen
