+++ app/src/main/java/com/example/learningapp/view/screen/FinishLessonsToCompete.kt
@@ -15,68 +15,59 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.constraintlayout.compose.ConstraintLayout
 import com.example.learningapp.R
 import com.example.learningapp.view.theme.Blue100
+import com.example.learningapp.view.theme.Orange100
 
 @Composable
 @Preview
-fun NoProfileScreen() {
+fun FinishLessonsToCompete() {
     ConstraintLayout(Modifier.fillMaxSize()) {
         val (nextButton, profileBox) = createRefs()
-        val email by remember { mutableStateOf("gmail address info") }
+        val lessonNum by remember { mutableStateOf("9 lessons") }
 
-        Button(
-            onClick = { },
-            colors = ButtonDefaults.buttonColors(Blue100),
-            modifier = Modifier
-                .padding(10.dp)
-                .fillMaxWidth()
-                .constrainAs(nextButton) {
-                    bottom.linkTo(parent.bottom)
-                    start.linkTo(parent.start)
-                    end.linkTo(parent.end)
-                }
-        ) {}
         Column(
             modifier = Modifier
                 .fillMaxWidth()
                 .constrainAs(profileBox) {
-                    bottom.linkTo(nextButton.top)
+                    bottom.linkTo(parent.bottom)
                     start.linkTo(parent.start)
                     end.linkTo(parent.end)
                     top.linkTo(parent.top)
                 },
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
+            Image(
+                painter = painterResource(id = R.drawable.leader_board_preview),
+                contentDescription = "Google Logo",
+                modifier = Modifier
+                    .size(300.dp)
+            )
             Text(
-                text = "There is no Duolingo profile for that google account",
+                text = buildAnnotatedString {
+                    append("Finish ")
+                    withStyle(style = SpanStyle(fontWeight = FontWeight.Bold, color = Orange100)) {
+                        append(lessonNum)
+                    }
+                    append(" to start competing Leaderboards")
+                },
                 Modifier
                     .padding(30.dp)
                     .fillMaxWidth(),
                 fontSize = 25.sp,
                 textAlign = TextAlign.Center,
                 fontWeight = FontWeight.Bold
-            )
-            Image(
-                painter = painterResource(id = R.drawable.google_search),
-                contentDescription = "Google Logo",
-                modifier = Modifier
-                    .size(80.dp)
-            )
-            Text(
-                text = email,
-                Modifier
-                    .padding(30.dp)
-                    .fillMaxWidth(),
-                fontSize = 20.sp,
-                textAlign = TextAlign.Center
             )
         }
     }
